add ability to view selected/hovered meshes
had to use mutable ref (useRef) to get around race condition with state updates + onPointerOut -- unfortunately was the case with either signals or react state hooks. Will need to smoothen this over to make the approach more consistent but for the moment things work and are performant for the purposes.